# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: FinDoc Analyzer API
  description: |
    API for FinDoc Analyzer application that processes financial documents, 
    extracts securities data, and provides portfolio analysis.
  contact:
    name: FinDoc Support
    url: https://findoc.example.com/support
    email: support@findoc.example.com
  version: 1.0.0
servers:
  - url: https://api.findoc.example.com/v1
    description: Production server
  - url: https://api-staging.findoc.example.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /documents/{documentId}/securities:
    get:
      tags:
        - documents
        - securities
      summary: Get securities
      description: Retrieve securities extracted from a document
      operationId: getSecurities
      parameters:
        - name: documentId
          in: path
          description: Document ID
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Securities data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  securities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Security'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Security:
      type: object
      properties:
        id:
          type: string
          example: sec-123
        name:
          type: string
          example: Apple Inc.
        symbol:
          type: string
          example: AAPL
        isin:
          type: string
          example: US0378331005
        assetClass:
          type: string
          example: Equity
        quantity:
          type: number
          format: double
          example: 100
        price:
          type: number
          format: double
          example: 175.25
        value:
          type: number
          format: double
          example: 17525.00
        currency:
          type: string
          example: USD
        sector:
          type: string
          example: Technology
        country:
          type: string
          example: US
        marketCap:
          type: string
          example: Large-Cap
        metadata:
          type: object
          additionalProperties: true
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Document not found
        code:
          type: string
          example: DOCUMENT_NOT_FOUND
        status:
          type: integer
          example: 404